Cleaned up some broken remotes
