+++ apps/shell/src/app/modules/settings/pages/settings-approvals/sub-pages/edit-program-approval/edit-program-approval.tsx
@@ -8,7 +8,7 @@ import {
   FormInputNumber,
 } from '@cased/ui';
 import { useCallback, useState } from 'react';
-import { ApprovalType, IApprovalProgram } from '@cased/remotes';
+import { ApprovalType, IApprovalProgram } from '@cased/data';
 import { useParams } from 'react-router-dom';
 import { useStoreActions } from '@cased/redux';
 import SettingsTemplate, { TabId } from '../../../../settings-template';
