+++ apps/shell/src/app/modules/settings/pages/settings-runbooks/databases/database-form/database-form.tsx
@@ -1,4 +1,4 @@
-import { IRunbookDatabase } from '@cased/remotes';
+import { IRunbookDatabase } from '@cased/data';
 import { Button, FormInputText, FormPartialHiddenPassword } from '@cased/ui';
 import { useCallback } from 'react';
