+++ apps/shell/src/app/modules/settings/pages/settings-users/tables/table-groups/table-groups.tsx
@@ -8,7 +8,7 @@ import {
   TableRow,
   TextLink,
 } from '@cased/ui';
-import { ISettingsGroup } from '@cased/remotes';
+import { ISettingsGroup } from '@cased/data';
 
 export interface TableGroupsProps {
   list: ISettingsGroup[];
