+++ apps/shell/src/app/modules/settings/pages/settings-users/tables/table-users-and-roles/table-users-and-roles.tsx
@@ -1,4 +1,4 @@
-import { ISettingsUser } from '@cased/remotes';
+import { ISettingsUser } from '@cased/data';
 import {
   Table,
   TableBody,
