+++ libs/data/src/lib/i-user.ts
@@ -1,0 +1,7 @@
+export interface IUser {
+  id?: string;
+  email?: string;
+  name?: string;
+  groups?: string[];
+  role?: 'admin' | 'user';
+}
