+++ libs/remotes/src/lib/activities/activities-service.ts
@@ -1,6 +1,6 @@
+import { ILog } from '@cased/data';
 import { saveAs } from 'file-saver';
 import { axiosInstance } from '../axios';
-import { ILog } from '../settings/i-settings';
 import { ISettingsLogsApiResponse } from '../settings/i-settings-api-response';
 import { transformLogs } from '../settings/settings-service.utilities';
