+++ libs/remotes/src/lib/remotes.tsx
@@ -4,10 +4,8 @@ export * from './axios/axios-service';
 export * from './runbooks/runbooks-service';
 export * from './storage/storage-service';
 export * from './settings/settings-service';
-export * from './settings/i-settings';
 export * from './prompt/prompt-service';
 export * from './activities/activities-service';
 export * from './settings/runbooks/settings-runbooks.service';
 export * from './settings/approvals/settings-approvals.service';
 export * from './users/users.service';
-export * from './settings/approvals/i-settings-approvals';
