+++ libs/remotes/src/lib/settings/runbooks/settings-runbooks.service.ts
@@ -1,29 +1,5 @@
+import { IApiProvider, IEntry, IRunbookDatabase } from '@cased/data';
 import { axiosInstance } from '../../axios';
-
-export interface IEntry {
-  id: string;
-  name: string;
-}
-
-export interface IApiProvider extends IEntry {
-  baseUrl?: string;
-  authType: 'basic' | 'token';
-  username?: string;
-  password?: string;
-  authToken?: string;
-  secretToken?: string;
-  hasPassword?: boolean;
-  hasToken?: boolean;
-}
-
-export interface IRunbookDatabase extends IEntry {
-  host?: string;
-  port?: string;
-  label?: string;
-  username?: string;
-  password?: string;
-  hasPassword?: boolean;
-}
 
 const getAllApiProviders = async (): Promise<IEntry[]> => {
   const {
