+++ libs/remotes/tsconfig.lib.json
@@ -19,5 +19,11 @@
     "**/*.spec.jsx",
     "**/*.test.jsx"
   ],
-  "include": ["**/*.js", "**/*.jsx", "**/*.ts", "**/*.tsx"]
+  "include": [
+    "**/*.js",
+    "**/*.jsx",
+    "**/*.ts",
+    "**/*.tsx",
+    "../data/src/lib/i-settings-approvals.ts"
+  ]
 }
